refactro: rename projects
