+++ GoodNews.API/GoodNews.API.csproj
@@ -19,9 +19,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\GoodNews.Core\GoodNews.Core.csproj" />
     <ProjectReference Include="..\GoodNews.Data\GoodNews.Data.csproj" />
     <ProjectReference Include="..\GoodNews.Migrations\GoodNews.Migrations.csproj" />
-    <ProjectReference Include="..\Services\Services.csproj" />
+    <ProjectReference Include="..\GoodNews.Services\GoodNews.Services.csproj" />
   </ItemGroup>
 
 </Project>
