+++ GoodNews.Core/GoodNews.Core.csproj
