+++ GoodNews.Core/IEmailSender.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
-using System.Threading.Tasks;
 
-namespace Core
+namespace GoodNews.Core
 {
     public interface IEmailSender
     {
