+++ GoodNews.Core/IRepository.cs
