+++ GoodNews.MVC/GoodNews.MVC.csproj
@@ -14,9 +14,9 @@
 
 
   <ItemGroup>
-    <ProjectReference Include="..\Core\Core.csproj" />
+    <ProjectReference Include="..\GoodNews.Core\GoodNews.Core.csproj" />
     <ProjectReference Include="..\GoodNews.Migrations\GoodNews.Migrations.csproj" />
-    <ProjectReference Include="..\Services\Services.csproj" />
+    <ProjectReference Include="..\GoodNews.Services\GoodNews.Services.csproj" />
   </ItemGroup>
