+++ GoodNews.Services/GoodNews.Services.csproj
@@ -12,7 +12,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Core\Core.csproj" />
+    <ProjectReference Include="..\GoodNews.Core\GoodNews.Core.csproj" />
+    <ProjectReference Include="..\GoodNews.CQS\GoodNews.CQS.csproj" />
     <ProjectReference Include="..\GoodNews.DAL\GoodNews.DAL.csproj" />
   </ItemGroup>
