+++ GoodNews.Services/Parsers/IS13Parser.cs
@@ -1,0 +1,8 @@
+﻿using GoodNews.Core;
+
+namespace GoodNews.Services.Parsers
+{
+    public interface IS13Parser : INewsParser
+    {
+    }
+}
