+++ GoodNews.Services/Parsers/ITutByParser.cs
@@ -1,0 +1,8 @@
+﻿using GoodNews.Core;
+
+namespace GoodNews.Services.Parsers
+{
+    public interface ITutByParser : INewsParser
+    {
+    }
+}
