+++ GoodNews.sln
@@ -3,11 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29324.140
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Services", "Services\Services.csproj", "{B90A43FC-64B5-4843-9792-EF371033566E}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GoodNews.DAL", "GoodNews.DAL\GoodNews.DAL.csproj", "{50F580ED-BB52-4DB3-BD88-A8DC79514354}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core", "Core\Core.csproj", "{CB43F48B-F9DF-4CEA-A1DC-762BE98B6BFF}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GoodNews.Data", "GoodNews.Data\GoodNews.Data.csproj", "{D467F67C-700B-4D01-8E50-B04A2ACF0648}"
 EndProject
@@ -17,7 +13,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GoodNews.API", "GoodNews.AP
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GoodNews.CQS", "GoodNews.CQS\GoodNews.CQS.csproj", "{D5416FA5-92B3-463C-B5FF-A48F74A9F0B0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoodNews.Migrations", "GoodNews.Migrations\GoodNews.Migrations.csproj", "{C800411D-802B-42FD-A2B5-730915ADC4B8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GoodNews.Migrations", "GoodNews.Migrations\GoodNews.Migrations.csproj", "{C800411D-802B-42FD-A2B5-730915ADC4B8}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GoodNews.Core", "GoodNews.Core\GoodNews.Core.csproj", "{FB060630-1CD4-4FE5-BBD8-5DC7BEC1269E}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GoodNews.Services", "GoodNews.Services\GoodNews.Services.csproj", "{B6146D4A-64AF-4A8E-A856-89EF7685A80A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,18 +25,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B90A43FC-64B5-4843-9792-EF371033566E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B90A43FC-64B5-4843-9792-EF371033566E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B90A43FC-64B5-4843-9792-EF371033566E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B90A43FC-64B5-4843-9792-EF371033566E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{50F580ED-BB52-4DB3-BD88-A8DC79514354}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{50F580ED-BB52-4DB3-BD88-A8DC79514354}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{50F580ED-BB52-4DB3-BD88-A8DC79514354}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{50F580ED-BB52-4DB3-BD88-A8DC79514354}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CB43F48B-F9DF-4CEA-A1DC-762BE98B6BFF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CB43F48B-F9DF-4CEA-A1DC-762BE98B6BFF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CB43F48B-F9DF-4CEA-A1DC-762BE98B6BFF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CB43F48B-F9DF-4CEA-A1DC-762BE98B6BFF}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D467F67C-700B-4D01-8E50-B04A2ACF0648}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D467F67C-700B-4D01-8E50-B04A2ACF0648}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D467F67C-700B-4D01-8E50-B04A2ACF0648}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -57,6 +49,14 @@ Global
 		{C800411D-802B-42FD-A2B5-730915ADC4B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C800411D-802B-42FD-A2B5-730915ADC4B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C800411D-802B-42FD-A2B5-730915ADC4B8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FB060630-1CD4-4FE5-BBD8-5DC7BEC1269E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FB060630-1CD4-4FE5-BBD8-5DC7BEC1269E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FB060630-1CD4-4FE5-BBD8-5DC7BEC1269E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FB060630-1CD4-4FE5-BBD8-5DC7BEC1269E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B6146D4A-64AF-4A8E-A856-89EF7685A80A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B6146D4A-64AF-4A8E-A856-89EF7685A80A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B6146D4A-64AF-4A8E-A856-89EF7685A80A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B6146D4A-64AF-4A8E-A856-89EF7685A80A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
