--- Services/Parsers/IS13Parser.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Core;
-
-namespace Services.Parsers
-{
-    public interface IS13Parser : INewsParser
-    {
-    }
-}
